backend/requirements.txt
# DesiFriend AI Backend Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# AI and Language Processing
openai==1.10.0
langdetect==1.0.9

# HTTP Requests
requests==2.31.0
httpx==0.26.0

# Audio Processing
pydub==0.25.1

# Environment Variables
python-dotenv==1.0.0
